+++ mirrord/agent/src/main.rs
@@ -4,7 +4,6 @@
 #![feature(is_some_and)]
 #![feature(let_chains)]
 #![feature(type_alias_impl_trait)]
-#![feature(unix_socket_abstract)]
 #![feature(tcp_quickack)]
 
 use std::{